Move theme-config to config folder.
